fix code smell
